+++ service-map/src/main/java/fi/mml/map/mapwindow/service/wms/WebMapServiceFactory.java
@@ -1,13 +1,16 @@
 package fi.mml.map.mapwindow.service.wms;
 
-import fi.mml.map.mapwindow.service.db.CapabilitiesCacheService;
-import fi.mml.map.mapwindow.service.db.CapabilitiesCacheServiceIbatisImpl;
+import fi.nls.oskari.domain.map.OskariLayer;
+import fi.nls.oskari.map.layer.OskariLayerService;
+import fi.nls.oskari.map.layer.OskariLayerServiceIbatisImpl;
+import fi.nls.oskari.service.OskariComponentManager;
+import fi.nls.oskari.service.capabilities.CapabilitiesCacheService;
 import fi.mml.map.mapwindow.util.RemoteServiceDownException;
 import fi.nls.oskari.cache.Cache;
 import fi.nls.oskari.cache.CacheManager;
-import fi.nls.oskari.domain.map.CapabilitiesCache;
 import fi.nls.oskari.log.LogFactory;
 import fi.nls.oskari.log.Logger;
+import fi.nls.oskari.service.capabilities.OskariLayerCapabilities;
 import fi.nls.oskari.wms.WMSCapabilities;
 
 /**
@@ -18,7 +21,8 @@ public class WebMapServiceFactory {
 
 	/** Logger */
 	private static Logger log = LogFactory.getLogger(WebMapServiceFactory.class);
-    private static final CapabilitiesCacheService capabilitiesCacheService = new CapabilitiesCacheServiceIbatisImpl();
+    private static final CapabilitiesCacheService capabilitiesCacheService = OskariComponentManager.getComponentOfType(CapabilitiesCacheService.class);
+    private static final OskariLayerService service = new OskariLayerServiceIbatisImpl();
     private static Cache<WebMapService> wmsCache = CacheManager.getCache(WebMapServiceFactory.class.getName());
     static {
         wmsCache.setExpiration(12L*60L*60L*1000L);
@@ -39,7 +43,8 @@ public class WebMapServiceFactory {
 		WebMapService wms = wmsCache.get(cacheKey);
         // caching since this is called whenever a layer JSON is created!!
 		if (wms == null) {
-            CapabilitiesCache cc = capabilitiesCacheService.find(layerId);
+            OskariLayer layer = service.find(layerId);
+            OskariLayerCapabilities cc = getCaps(layer);
             if(cc == null) {
                 // setup empty capabilities so we don't try to parse again before cache flush
                 WMSCapabilities emptyCaps = new WMSCapabilities();
@@ -47,10 +52,11 @@ public class WebMapServiceFactory {
                 return emptyCaps;
             }
             try {
-                if (cc != null && "1.3.0".equals(cc.getVersion().trim())) {
-                    wms = new WebMapServiceV1_3_0_Impl("from DataBase", cc.getData().trim(), layerName);
-                } else if (cc != null && "1.1.1".equals(cc.getVersion().trim())) {
-                    wms = new WebMapServiceV1_1_1_Impl("from DataBase", cc.getData().trim(), layerName);
+                final String data = cc.getData().trim();
+                if (isVersion1_3_0(data)) {
+                    wms = new WebMapServiceV1_3_0_Impl("from DataBase", data, layerName);
+                } else if (isVersion1_1_1(data)) {
+                    wms = new WebMapServiceV1_1_1_Impl("from DataBase", data, layerName);
                 }
                 // cache the parsed value
                 wmsCache.put(cacheKey, wms);
@@ -62,6 +68,14 @@ public class WebMapServiceFactory {
 		}
 		return wms;
 	}
+
+    private static OskariLayerCapabilities getCaps(OskariLayer layer) throws WebMapServiceParseException {
+        try {
+            return capabilitiesCacheService.getCapabilities(layer);
+        } catch (Exception ex) {
+            throw new WebMapServiceParseException(ex);
+        }
+    }
 
     public static void flushCache(final int layerId) {
         wmsCache.remove("wmsCache_"+layerId);
@@ -78,10 +92,10 @@ public class WebMapServiceFactory {
 	 * @return
 	 */
 	public static boolean isVersion1_1_1(String data) {
-		if (data.indexOf("WMT_MS_Capabilities version=\"1.1.1\"") > 0) {
+		if (data.contains("WMT_MS_Capabilities version=\"1.1.1\"")) {
 			return true;
 		} else {
-            return data.indexOf("WMT_MS_Capabilities updateSequence=\"1\" version=\"1.1.1\"") > 0;
+            return data.contains("WMT_MS_Capabilities updateSequence=\"1\" version=\"1.1.1\"");
         }
 	}
 
@@ -92,7 +106,7 @@ public class WebMapServiceFactory {
 	 * @return
 	 */
 	public static boolean isVersion1_3_0(String data) {
-		return data.indexOf("WMS_Capabilities version=\"1.3.0\"") > 0;
+		return data.contains("WMS_Capabilities version=\"1.3.0\"");
 	}